Lock OS thread for main goroutine
In short if the main code performs network operations assuming
they are done in the default namespace this may not always be the case.
locking the goroutine to the current OS thread ensures that any
operation that is performed, is performed in the default namespace.

While i have not encounterated a case where an operation ran
in the wrong namespace as all operations that are performed in a
namespace other than the default one are performed after locking the OS
thread to the goroutine (and released after).

However, i see the same pattern is done in other CNI's...